Replace JSON-CDC with CCF for events
Also moved test function SDKEventToFlow to unittest package.
