+++ convert/flow.go
@@ -24,7 +24,8 @@ import (
 	"github.com/onflow/flow-emulator/types"
 
 	"github.com/onflow/cadence"
-	jsoncdc "github.com/onflow/cadence/encoding/json"
+	"github.com/onflow/cadence/encoding/ccf"
+
 	sdk "github.com/onflow/flow-go-sdk"
 	sdkcrypto "github.com/onflow/flow-go-sdk/crypto"
 	"github.com/onflow/flow-go/access"
@@ -182,23 +183,8 @@ func FlowTransactionResultToSDK(result *access.TransactionResult) (*sdk.Transact
 	return sdkResult, nil
 }
 
-func SDKEventToFlow(event sdk.Event) (flowgo.Event, error) {
-	payload, err := jsoncdc.Encode(event.Value)
-	if err != nil {
-		return flowgo.Event{}, err
-	}
-
-	return flowgo.Event{
-		Type:             flowgo.EventType(event.Type),
-		TransactionID:    SDKIdentifierToFlow(event.TransactionID),
-		TransactionIndex: uint32(event.TransactionIndex),
-		EventIndex:       uint32(event.EventIndex),
-		Payload:          payload,
-	}, nil
-}
-
 func FlowEventToSDK(flowEvent flowgo.Event) (sdk.Event, error) {
-	cadenceValue, err := jsoncdc.Decode(nil, flowEvent.Payload)
+	cadenceValue, err := ccf.Decode(nil, flowEvent.Payload)
 	if err != nil {
 		return sdk.Event{}, err
 	}
