+++ storage/encoding_test.go
@@ -21,8 +21,6 @@ package storage
 import (
 	"testing"
 
-	"github.com/onflow/flow-emulator/convert"
-
 	"github.com/onflow/flow-go-sdk/test"
 	flowgo "github.com/onflow/flow-go/model/flow"
 	"github.com/stretchr/testify/assert"
@@ -116,8 +114,8 @@ func TestEncodeEvents(t *testing.T) {
 
 	t.Parallel()
 
-	event1, _ := convert.SDKEventToFlow(test.EventGenerator().New())
-	event2, _ := convert.SDKEventToFlow(test.EventGenerator().New())
+	event1, _ := unittest.SDKEventToFlow(test.EventGenerator().New())
+	event2, _ := unittest.SDKEventToFlow(test.EventGenerator().New())
 
 	events := []flowgo.Event{
 		event1,
